Remove all dependencies.
I have thought about ORM bullshit a lot and realized that this library
should be code-generation-only.
We will provide a fuck ton of utility code generators and maybe
eventually migration functionality I had initially in mind when I was
starting XSQL, but for now we'll just stick to the bare minimum because
it's bloody impossible to implement somewhat complete SQL transcoder in
any sane amount of time. And the efforts from writers of all sorts of
ORMs are laughable.